+++ test/mocha/10-http.js
@@ -1,12 +1,12 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const bedrock = require('bedrock');
+import * as bedrock from '@bedrock/core';
+import * as helpers from './helpers.js';
+import {createRequire} from 'module';
+import {agent} from '@bedrock/https-agent';
+const require = createRequire(import.meta.url);
 const {CapabilityAgent} = require('@digitalbazaar/webkms-client');
-const helpers = require('./helpers');
-const {agent} = require('bedrock-https-agent');
 const {httpClient} = require('@digitalbazaar/http-client');
 
 describe('bedrock-service-core HTTP API', () => {