+++ test/mocha/mock.data.js
@@ -1,18 +1,15 @@
 /*!
  * Copyright (c) 2018-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import {config} from '@bedrock/core';
 
-const {config} = require('bedrock');
-
-const data = {};
-module.exports = data;
+export const mockData = {};
 
 // mock product IDs and reverse lookup for service products
-data.productIdMap = new Map([
+mockData.productIdMap = new Map([
   // example service
   ['example', 'urn:uuid:66aad4d0-8ac1-11ec-856f-10bf48838a41'],
   ['urn:uuid:66aad4d0-8ac1-11ec-856f-10bf48838a41', 'example']
 ]);
 
-data.baseUrl = config.server.baseUri;
+mockData.baseUrl = config.server.baseUri;