+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # bedrock-service-core ChangeLog
 
+## 4.0.0 - 2022-04-xx
+
+### Changed
+- **BREAKING**: Rename package to `@bedrock/service-core`.
+- **BREAKING**: Convert to module (ESM).
+- **BREAKING**: Remove default export.
+- **BREAKING**: Require node 14.x.
+
 ## 3.1.1 - 2022-03-28
 
 ### Fixed